[Silver IV] Title: 수 찾기, Time: 580 ms, Memory: 52184 KB -BaekjoonHub

+++ 백준/Silver/1920. 수 찾기/수 찾기.java
@@ -1,10 +1,8 @@
-
+// 메모리 53840
+// 시간   580
 
 import java.io.*;
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Map;
-import java.util.Scanner;
 import java.util.Set;
 
 public class Main {
@@ -14,8 +12,8 @@ public class Main {
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
         // 1번째 라인의 입력 M을 받는다.
-        long N = Integer.parseInt(br.readLine());
-        Set<Long> A = new HashSet<>((int) N);
+        int N = Integer.parseInt(br.readLine());
+        Set<Long> A = new HashSet<>(N);
         String[] strArr;
 
         // br로 부터 입력 받는 '4 1 5 2 3 문자열'크기5의 정수 배열로 변환한다.
@@ -25,7 +23,7 @@ public class Main {
         }
 
         // 3번째 라인의 입력 M을 받는다.
-        long M = Long.parseLong(br.readLine());
+        int M = Integer.parseInt(br.readLine());
         strArr = br.readLine().split(" ");
         for (int i = 0; i < M; i++) {
             if(A.contains(Long.parseLong(strArr[i]))){
